# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified Communications Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountProvider } from "./accountProvider";

export interface Account {
  id: number;
  email: string;
  name: string;
  provider: AccountProvider;
  color: string;
  isActive: boolean;
  unreadCount: number;
  createdAt: string;
}