# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified Communications Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  messageCount: number;
  /** @nullable */
  lastMessageAt?: string | null;
  createdAt: string;
}