# lib/api-zod/src/generated/types/createAttachmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified Communications Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAttachmentBody {
  filename: string;
  mimeType: string;
  size: number;
  /** @nullable */
  url?: string | null;
}