# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified Communications Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Attachment } from "./attachment";

export interface Message {
  id: number;
  accountId: number;
  accountEmail: string;
  accountName: string;
  accountColor: string;
  folder: string;
  subject: string;
  fromName: string;
  fromEmail: string;
  toList: string;
  /** @nullable */
  ccList?: string | null;
  /** @nullable */
  bodyText?: string | null;
  /** @nullable */
  bodyHtml?: string | null;
  isRead: boolean;
  isStarred: boolean;
  hasAttachments: boolean;
  attachments: Attachment[];
  receivedAt: string;
  createdAt: string;
}