# lib/api-zod/src/generated/types/messagesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified Communications Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";

export interface MessagesResponse {
  messages: Message[];
  total: number;
  hasMore: boolean;
}