# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified Communications Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from "./contact";
import type { Message } from "./message";

export interface SearchResults {
  messages: Message[];
  contacts: Contact[];
  query: string;
  totalMessages: number;
  totalContacts: number;
}