# lib/api-zod/src/generated/types/updateContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified Communications Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateContactBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  notes?: string | null;
}